fix clerk auth in client projects route and wrap client dashboard
